+++ project/src/main/java/com/company/banking/observer/Observer.java
@@ -1,9 +1,7 @@
 package com.company.banking.observer;
 
-import com.company.banking.domain.Account;
-import com.company.banking.domain.bankaccount.CompanyAccount;
-import com.company.banking.domain.bankaccount.PersonalAccount;
-import com.company.banking.domain.AccountEntry;
+import com.company.framework.domain.Account;
+import com.company.framework.domain.AccountEntry;
 
 public interface Observer {
     void update(Account account, AccountEntry accountEntry);
