+++ project/src/main/java/com/company/banking/service/BankingService.java
@@ -1,15 +1,12 @@
 package com.company.banking.service;
 
-import com.company.banking.domain.Account;
+import com.company.framework.domain.Account;
+import com.company.framework.service.AccountService;
 
 import java.util.Collection;
 
-public interface BankingService {
-        void createAccount(Account account);
-        Account getAccount(String accountNumber);
-        Collection<Account> getAllAccounts();
-        void deposit (String accountNumber, double amount);
-        void withdraw (String accountNumber, double amount);
+public interface BankingService extends AccountService{
+
         void transferFunds(String fromAccountNumber, String toAccountNumber, double amount, String description);
         void addInterest();
         void generateAccountReports();
