+++ project/src/main/java/com/company/framework/service/AccountService.java
@@ -1,0 +1,13 @@
+package com.company.framework.service;
+
+import com.company.framework.domain.Account;
+
+import java.util.Collection;
+
+public interface AccountService {
+    void createAccount(Account account);
+    Account getAccount(String accountNumber);
+    Collection<Account> getAllAccounts();
+    void deposit (String accountNumber, double amount);
+    void withdraw (String accountNumber, double amount, String description);
+}
